Discovering Room UI with the ability to collapse video div when hang up (#56)
User can now enter room and view video call on top with a ratio of 9:7 video and room selection ratio. I took this ratio from YouTube theater mode.

I used CustomEvent to avoid needing a global Stimulus controller, not sure if this is the 
stimulus way but it feel right for now.
I read https://www.betterstimulus.com/ to learn writing Stimulus in this pull request.

TODO: Extracting an adapter module for interacting with Jitsi Meet in the "Convene" way.