Install Rspec for openbox testing of convene-web
Rspec is a nice complement to Cucumber, in that it allows us to directly
integrate with the underlying Ruby and Rails code without writing human
or computer interfaces that require inter-process communication.

This helps us write small, focused tests that deal with 1~3 Ruby classes
instead of having to spin up a working instance of the entire
application.

Tests that need database access should `require "rails-helper"`, and
tests that can be executed without a database should `require "spec-helper"`"