Seeds are idempotent and fail with an exception (#69)
I wound up having duplicated workspaces because the name and slug was
not unique after running the seeds multiple times.  Now running
`db:seed` should not create the same workspace or room on additional
runs, and it should also raise exceptions if it cannot perform an update
or create operation.