Streamline running convene by adding docs + helper scripts
We noticed that there was not any way to know how to get started running
the Convene::Web interface unless you happened to know how to explore
the project repository already.

By sprouting a `bin/run` script that launches the rails server, as well
as documenting how to get started in the project root readme with more
technical details in the Convene::Web README we think this will make it
easier for people to pick up the project and start contributing.

H/t @KellyAH for helping figure out where to put what documentation
updates!