Wire in scenarios for guests not discovering unlisted rooms
See: https://github.com/zinc-collective/convene/issues/72
See: https://github.com/zinc-collective/convene/issues/39

- You can bypass waiting for a room card to exist by passing `false` to
  the findRoomCard method on workspace

We think that it would be nice to make the `findRoomCard` function to
return the roomCard itself, or return a false; and move the wait out of
it completely; which would push responsibility for knowing whether we
should "wait for it" to the step.

This would also allow us to "check" the value of the roomCard in steps.

- We use `cucumber-pretty` to format our output legibly

Co-authored-by: Tom Lee <user512@users.noreply.github.com>
Co-authored-by: CJ Joulain <joulain@gmail.com>
Co-authored-by: KellyAH <KellyAH@users.noreply.github.com>