🪒🧹 Locking rooms feature definition is more expressive
See: https://github.com/zinc-collective/convene/issues/40

We noticed that there was ~8 ish scenarios, each of which could be
reduced to a single `then` statement, where we authenticate as the
distinct actor-types and then perform the action.

This also makes it easier for someone who is trying to figure out what
the rules are for entering a locked room because it's not spread out
across 100 lines of feature definition.