🔧 Steps may match unnamed Rooms with the Room parameter type
This through us for a bit of a loop, because we didn't understand how
parameterTypes got their arguments for their Transformers. We did some
digging into the underlying Cucumber-JS code-base; to figure out what
may be happening; and discovered that paramterType matching discards the
full content of the match in cases where there are child matches.