📝Sketch out step definitions for locking rooms features
As we were writing the step definitions, we noticed a few things like
`a` vs `an` having computational meaning despite not having semantic
meaning. We loosened up the paramaterTypes so that `a` or `an` are
acceptable, and we introduced a `roomKey` parameter type for matching
steps based upon right or wrong room keys.