Begin to wire in feature test for Locking Rooms (#111)
* 🪒🧹 Locking rooms feature definition is more expressive

See: https://github.com/zinc-collective/convene/issues/40

We noticed that there was ~8 ish scenarios, each of which could be
reduced to a single `then` statement, where we authenticate as the
distinct actor-types and then perform the action.

This also makes it easier for someone who is trying to figure out what
the rules are for entering a locked room because it's not spread out
across 100 lines of feature definition.

* 🗝🚪✅ Lay out scenarios for locking and unlocking rooms

Also, add links to the wireframes so that folks who are reading
the scenarios have an idea what we are working towards

* 🔧 Steps may match unnamed Rooms with the Room parameter type

This through us for a bit of a loop, because we didn't understand how
parameterTypes got their arguments for their Transformers. We did some
digging into the underlying Cucumber-JS code-base; to figure out what
may be happening; and discovered that paramterType matching discards the
full content of the match in cases where there are child matches.

* 📝Sketch out step definitions for locking rooms features

As we were writing the step definitions, we noticed a few things like
`a` vs `an` having computational meaning despite not having semantic
meaning. We loosened up the paramaterTypes so that `a` or `an` are
acceptable, and we introduced a `roomKey` parameter type for matching
steps based upon right or wrong room keys.

* 🤭🐞 The expanded search made some of the other tests fail

Turns out, we had not properly accounted for the expanded match
including `"` and other characters in the room name! Oops :X.

Co-authored-by: CJ Joulain <joulain@gmail.com>