Add table of contents and summary sentance (#124)
Because it really bothered me that when I first read Convene's readme
none of the text clearly told me what kind of app Convene was. And a TOC
makes it easier for readers to jump to the parts they care about and
when the readme starts getting really long, a TOC really helps.