Sprout UI for Locking and Unlocking the Room
There's a bit more to do here, for instance we want to
make sure that only people who have the existing room key
can unlock the room or change the room key.

This can probably be done by re-using some of the `waiting room`?

Or should we start pulling in pundit and using that for permissioning
before we wind up rolling our own?

TODO:
- [ ] Make the Access Code label turn red when there is an error
- [ ] Block access to the edit/update routes unless the room is
  enterable by the actor. (I.e access key is the correct key for the room)
- [ ] Pull out the Tailwind CSS classes into our semantic user
- [ ] Figure out where to put the "Settings" gear for the room (Room
  Card?)