Add information about zinc collective to readme and reorged info (#141)
The readme contained organization and voting info about The Zinc Collective
but IMHO it did not clearly explain what the org is. So I added that
info and moved all the organization info to the bottom. And I moved all
the budget and purpose sections having to do with the convene app to an
appropriate section.