Rough draft for decoupling our tests from room names
See: https://github.com/zinc-collective/convene/issues/40

As we implemented the story tests for Locked rooms, we started to notice
a lot of places where we were hardcoding strings that made assumptions
about what the database would have in it. This is a stab at shifting
responsibility for knowing how to locate Rooms and RoomCards based upon
the Room's particular attributes (locked, listed, etc) instead of
hard-coded names.