Wrap Webdriver behind Components/Pages
A lot of the logic for how to find and interact with the application has
been codified directly into step definitions. This required those of us
writing step definitions to know a lot about how webdriver works at a
low level.

Instead, this shifts us to a `Component` model, where Step Definitions
interact with the application through an Adapter that knows more about
how the underlying page works; and allows us to write steps that are a
bit more focused on "who is doing what now?" and less about "this is how
I find a particular element, then this is how I interact with it"