Switch to Hivemind, remove NPM in favor of Yarn
Hivemind has fewer dependencies than Overmind, which makes it a bit
easier for folks to get set up with.

And we noticed that we were using yarn in convene-web, which seems odd
that we're using different package managers for node in different parts
of the project; given how small the project is.