Use pessimistic version constraints in Gemfile
When we do not specify the version, we run the risk that a `bundle
update` will update package(s) in a way that is incompatible. If we use
the twiddle-wakka `~>` syntax, we are less likely to run into an incidental
incompatibility.

There is something to be said for leaving dependecy resolution unbounded
and expecting that our test suite will catch incompatibilities; but I'm
not sure I'm there yet.

see:
- https://guides.rubygems.org/patterns/#pessimistic-version-constraint
- https://thoughtbot.com/blog/rubys-pessimistic-operator