Flesh out Space Invitation RSVP request spec
We decided that for Guests, we want to allow them to bundle their
creation of a Person record with their response to the Invitation; so
that they don't need to log in before RSVPing.

However, we may want to think through what happens if they can't confirm
the email address they are claiming.

There's some more thinking we need to do around making sure that people
who invite someone and then enter a different email address can't just
claim invitations on others behalf.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>