Prevent the flow from exploding, but wants more...
We _think_ what we want to do, is if an Invitation is sent to an email
address we already have on file; we don't even show them the ability to
accept the email address unless they are signed in.

This way someone can't just "accept" invitations for other people.

However, we didn't have time to rework the RSVP show flow _and_ fix the
bug where it explodes horribly, so we defered the improvements to the
flow (for now)

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>