Simplify authoring of Parameter Types
Previously, we were including article ajdectives in our paramater type
matches. This was convenient, because we could write Gherkin where "an"
or "the" or "a" was interchangable, hooray!

However, it was also a bit of a pain because each parameter type wound
up wanting these article adjectives, which meant more complex regexes.
And the more complex regex's meant translating between Gherkin and step
definition became more difficult as well.

This introduces an `a` paramater type; so that steps can Gherkin can be
written with either `a` or `an` or `the` and still get matched to the
appropriate step definition; but without requiring each parameter type
to include it.

Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>