"Fix" race condition when checking email
We noticed that if the tests were running for the first time, we would
sometimes try to check the email was delivered before the background job
had completed; which would cause tests to fail sporadically.

Instead, we're doing a little retry thingy, where if no emails are
found, we delay 10ms, and try again. We might want to bump this up to
100ms but we'll see.

Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>