Shift responsibility from Invitation to SpaceEditPage
It didn't quite make sense for our domain object to know about where in
the UI to check it's status. Maybe if we had an API where we could just
get the invitation id and confirm it's status, it would make sense for
this to live on the `Invitation` class, but... we ain't there.

Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>