Automate check for Space Owner inviting a new space member (#379)
- Authorize Space when Editing Space
Previously, we wre relying on the `current_space` method to do
authorization, but that always uses `show?` which means anyone can edit
a Space, regardless of their rights!



Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>


- Added Invitation steps and parameter types
- Actor and Space parameter types now accept article adjectives
- SpaceEditPage has an `invite` function
- SystemTestSpace seeds a `space-owner@example.com` member
- Clean up some invalid assumptions when parsing an actor
- Invitation emails can find their URL

Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Tom lee <User512@users.noreply.github.com>

- "Fix" race condition when checking email

We noticed that if the tests were running for the first time, we would
sometimes try to check the email was delivered before the background job
had completed; which would cause tests to fail sporadically.

Instead, we're doing a little retry thingy, where if no emails are
found, we delay 10ms, and try again. We might want to bump this up to
100ms but we'll see.

Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>

- Assert that Invitation Status is visible on Space Edit Page

- added `Page#hasContent` for asserting against copy
- made the `Invitation` paramter type less greedy

Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>

- Shift responsibility from Invitation to SpaceEditPage


It didn't quite make sense for our domain object to know about where in
the UI to check it's status. Maybe if we had an API where we could just
get the invitation id and confirm it's status, it would make sense for
this to live on the `Invitation` class, but... we ain't there.

Co-authored-by: CJ Joulain <cjoulain@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Tom lee <User512@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>