Run PRs against the code in the branch, not Development
Essentially, we were running our PRs using the development code, not the code
in the branch. This was causing CI to "pass" in cases where development was
green, but the PR was red; and vice versa.

Now CI runs using the code in the branch, so hopefully our attempts to fix it
will actually run against our new code.

See: https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows#pull_request_target

Co-authored-by: Tom Lee <User512@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>