Payments are mapped back to the Account the Payer Selected
Turns out, the reason the Check wasn't getting saved was because we were
trying to validate a field before we had exchanged the public token for
the access token and item id.

However, we also learned that the item_id is insufficient, because it
returns all the accounts the person has within that institution; not
just the particular account that they authorized us to withdraw funds
from. Big scary. Potentially risky API design we'll need to design
around.

We now store a Account Description, as well as the Plaid Account ID with
the Check, so that we can map it back both "visually" and by comparing
the Account IDs returned in other Plaid API calls

Co-authored-by: Ana Ulin <ana@ulin.org>