Eliminate race condition when sending multiple emails to same person
Turns out, because we send the Invitation to the Space and the Space
Passcode to the same person; and the timing of delivery could be slower
than the check for the email, we were acting as if the Invitation to the
Space _was_ the email with the Space Passcode.

Instead, we now have the ability to filter our Email queries based upon
the text of the email; so that we can wait until that condition is also
fulfilled, instead of assuming the most recent email received is the
email we wanted it to be.

Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>