Indicate that The `a` and `a2` params are unused
Step definitions _must_ match the Arity, so our Filler parameter types
(like `a`) that provide flexibility at the Gherkin level must be
represented in the step definition itself.

The following Step uages, for example, all map to the same step
definition:
```gherkin
When the Invitation to "a-guest@example.com" for the "System Test" Space
  is accepted by the Guest "a-guest@example.com"

When an Invitation to "a-guest@example.com" for a "System Test" Space is
  accepted by a Guest "a-guest@example.com"
```

but the swapping of `an` and `the` are programmaticaly irrelevant.

Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>