Add Validation to Checks (#463)
* Begin adding Validations to Payment Form

- Redo the form to use the text_field and error partials
- Unwind some of the over-specification-ishness of the CSS
- Consolidate "field" or "attribute" in component partials to
  "attribute"

* Payments are mapped back to the Account the Payer Selected

Turns out, the reason the Check wasn't getting saved was because we were
trying to validate a field before we had exchanged the public token for
the access token and item id.

However, we also learned that the item_id is insufficient, because it
returns all the accounts the person has within that institution; not
just the particular account that they authorized us to withdraw funds
from. Big scary. Potentially risky API design we'll need to design
around.

We now store a Account Description, as well as the Plaid Account ID with
the Check, so that we can map it back both "visually" and by comparing
the Account IDs returned in other Plaid API calls


* Don't let plaid hidden fields get "autocompleted"

Turns out, Firefox sometimes replaces the values of hidden inputs with
the values they had after the form was submitted. This can be scary; so
we are telling Firefox plz do not autocomplete these hidden fields; so
that Plaid can fill them out.

* Require a positive amount

* Consolidate the check dropbox's  `checks/new` and `in_room` views

* Provide a default Item type to ItemRepository, and fix the spec.

* Use VCR to fix the Checks request spec.

- Add the VCR gem and configure it.
- Use it to record interactions with the Plaid API in the Checks spec.

* Improve error message when Check bank account is missing.
* Add validation for email format and specs for all validations.


Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Ana Ulin <ana@ulin.org>