Rails-ujs is probably safe to bump to 7
In https://github.com/zinc-collective/convene/pull/495, we thought we
didn't want to jump from 6 to 7 without bumping Rails first.

We also thought that maybe we didn't even need UJS, because we have no
use-cases where we are explicitely using it.

However, after removing UJS
https://github.com/zinc-collective/convene/pull/576 we discovered that
we _do_ use it for signing people out!

We did a review of the changelog and the commits between 6.X and 7.0 for
ujs, and they seemed computationally meaningless; so we think it's safe
to just bump and defer the deletion.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>