Invitations include name _or_ email of the Invitor
See: https://github.com/zinc-collective/convene/issues/117#issuecomment-962519714

@anaulin had noticed that the invitation emails were left dangling after
"Invited by..." when the Invitor doesn't have a name.

We've pulled out a `display_name` method for the Person, and delegate it
explicitely in the `Invitation` as the `invitor_display_name`.

This gives us a seam for making more intelligent decisions about how we
present peoples names/emails/identities/etc in the application, but does
not exercise that seam to aggressively.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>