Use the branch of rspec-rails that supports Rails 7
See: https://github.com/rspec/rspec-rails/pull/2576

Basically, turns out there _is_ a version of rspec-rails that works with
Rails 7, however it is living in a branch atm because it's a big change.
(totally reasonable).

This uses a forked version of that branch, because it expects to be a
prerelease version, and I wasn't sure how to move my Gemfile to use
pre-release versions of itself, and once I started going down that path
I got skared.