Reframe to a Specific Example of what Blueprints do
This makes it a little bit more clear what a Blueprint is for, by giving
an example of the _output_ of the Blueprint. We still don't know what
the shape of the Blueprint data probably wants to look like, but from a
product-engineering perspective this is probably a decent step.

Co-authored-by: Ana Ulin <anaulin@users.noreply.github.com>