Sketch out the initial feature file for Space Blueprints (#568)
See: https://github.com/zinc-collective/convene/issues/476

Ana had noticed that our Blueprint's overwrite stuff they probably
shouldn't; and I had noticed something similar previously as well. I had
initially tossed Blueprints into the codebase as an affordance for
developers to quickly create a Space with initial data.

However, I'm beginning to think that it makes sense to be a bit more
intentional with them, and am promoting them to a sub-feature of the
Space feature.

This makes it a little bit more clear what a Blueprint is for, by giving
an example of the _output_ of the Blueprint. We still don't know what
the shape of the Blueprint data probably wants to look like, but from a
product-engineering perspective this is probably a decent step.

Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Ana Ulin <anaulin@users.noreply.github.com>