Refactor: Migrate responding to Invitations to it's own Feature file
We noticed that we were rapidly approaching the 3+ scenarios mark for
how People can respond to Invitations, which implied that we want an
additional feature file _just_ for that particular use-case.

We pulled that out, and in a fit of 'let's have fun with it!' started
reframing from "system test" to "Team Avatar" and that uncovered missing
stairs in how we interact with the test data.

At present, most (if not all) of our Test Data is "seeded" outside of
the test suite; which means that as scenarios play, they can leave the
data in a spot that doesn't work for future scenarios.

It also means that we can't leverage the affordances for differentiating
between Spaces, Rooms and Actors etc. using their names.

So now we're going to start driving out APIs for creating Space
Memberships and Authentication Methods so our Scenarios can set their
data up themselves.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>