Sprout API Client for AuthenticationMethod API
As we were driving out the test implementation, we noticed there is not
yet an API client method for creating AuthenticationMethods.

So we started to make one! Our API class is getting to be a bit more
fully featured, and our JavaScript "models" are as well; which is both
exciting and terrifying, because it means we're likely going to start
noticing a tension between what those Models want to do as "API Models"
vs as "Test" objects.

But I think we should be able to navigate that!

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>