WIP: Begin to persist Parameterized data between Steps
We noticed that when we were trying to runt he same test over again it
was giving us an error that the authentication method already exists.

To get around that, we're starting to share state between steps,
specifically so we can generate a random differentiator when creating
data in subsequent test runs.

This preserves our "append-only" feature test strategy, while still
giving us affordances for minimizing cross-talk between scenarios.

It's a bit messy, and there's likely a few more things to do, in
particular around serializing paramterized data before and after sending
it across the wire to the Application code.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>