Sprouts API for Creating Authentication Methods
The AuthenticationMethod API wants to be used in the feature tests to
make it a little bit easier to create People and authenticate as them
programmatically.

This pulls the work done in
https://github.com/zinc-collective/convene/pull/594 for exposing that
API out into it's own PR that we can merge independently, then rebase
into the feature branch for Invitations.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>