Creating a Space via the API doesn't Assume the SystemTest Blueprint
As I was working on making it possible to create arbitrary Spaces via
Cucumber in https://github.com/zinc-collective/convene/pull/644, I ran
into a bump because the API assumes you want to always create a
SystemTest space.

But that's still way too big of a step to do in a single PR, so I pulled
out a smaller step towards that larger goal.

Right now, the SpacesController#create is more of a nupsert than a true
Create, in that it:

1. Finds or Creates the Space based upon it's name
2. But if it _does_ exist, it does nothing. Hence Nupsert. Kinda like a
   no-op upsert.