Collect responsibility for applying a Space::Blueprint
OK, so I agree that the callback system was probably a little fraught.

Complexity emerges at those seams.

It may make sense for us to pull the responsibility into a factory.

Now if we want to create a space without executing a pile of callbacks,
we can.

And if we want to keep the controllers focused on the secure translation
of the HTTP protocol into domain objects we can.

And it sides steps a `Service` class (which I find reminiscent of the
`Manager` classes of my JavaDays.