Sprout Catering Pop-Up Furniture Feature Definition
This describes the two main scenarios, opening up a Pop-Up, and Placing
an Order at the Pop-Up.

There's more that could be built (in particular providing Space Owner's
with the ability to upload a Menu) but if we start here it should be
possible to throw together something that could work for Piikup as a
super-tiny proof of concept.