Locking and unlocking rooms feature uses explicit test data
We had been using implicit inferences about which space to use; which
was kind of ok but not really; because it would leave test data in a
strange state after the tests had been run, requiring unwinding the
changes semi-manually.

This moves the Locking and Unlocking Rooms feature definition to use our
more explicit test data setup. This should reduce brittleness when
making broader changes to our tests.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>