Invitations: Redesign Invitations (#594)
See: https://github.com/zinc-collective/convene/issues/117

There were some strange behaviors with how Invitations work, in particular:

- We never confirmed that Guests had access to the email address; we just assumed if they had the link they were totally the right person. That's not a good idea; so we made it necessary for them to confirm their access to the email address by using a OneTimePassword
- Neighbors couldn't accept Invitations at all; which seemed not-great.
- Once an Invitation was accepted, the user was just kinda left hanging.

This resolves those rough edges, as well as fleshes out the full feature test suite for inviting people to a Space.

* Refactor: Migrate responding to Invitations to it's own Feature file

We noticed that we were rapidly approaching the 3+ scenarios mark for
how People can respond to Invitations, which implied that we want an
additional feature file _just_ for that particular use-case.

We pulled that out, and in a fit of 'let's have fun with it!' started
reframing from "system test" to "Team Avatar" and that uncovered missing
stairs in how we interact with the test data.

At present, most (if not all) of our Test Data is "seeded" outside of
the test suite; which means that as scenarios play, they can leave the
data in a spot that doesn't work for future scenarios.

It also means that we can't leverage the affordances for differentiating
between Spaces, Rooms and Actors etc. using their names.

So now we're going to start driving out APIs for creating Space
Memberships and Authentication Methods so our Scenarios can set their
data up themselves.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>

* Sprout API Client for AuthenticationMethod API

As we were driving out the test implementation, we noticed there is not
yet an API client method for creating AuthenticationMethods.

So we started to make one! Our API class is getting to be a bit more
fully featured, and our JavaScript "models" are as well; which is both
exciting and terrifying, because it means we're likely going to start
noticing a tension between what those Models want to do as "API Models"
vs as "Test" objects.

But I think we should be able to navigate that!

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>

* Persist Parameterized data between Steps

We noticed that when we were trying to runt he same test over again it
was giving us an error that the authentication method already exists.

To get around that, we're starting to share state between steps,
specifically so we can generate a random differentiator when creating
data in subsequent test runs.

This preserves our "append-only" feature test strategy, while still
giving us affordances for minimizing cross-talk between scenarios.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>

* Complete Invitation only when confirmed Authentication Method matches

Previously, we confirmed Invitations prior to confirming the
authentication method; which... doesn't seem quite right.

Now, we have an interstitial that prevents invitations from completing
unless the AuthenticationMethod is confirmed, and if it is not it sends
the authentication code so that the person responding to the invitation
can confirm they are who they are

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>



* Locking and unlocking rooms feature uses explicit test data

We had been using implicit inferences about which space to use; which
was kind of ok but not really; because it would leave test data in a
strange state after the tests had been run, requiring unwinding the
changes semi-manually.

This moves the Locking and Unlocking Rooms feature definition to use our
more explicit test data setup. This should reduce brittleness when
making broader changes to our tests.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>

* Don't upsert test-id into Guest email addresses

Guests aren't real! I mean, they're real; but they're not tied to a
record in the database, so assuming we need to inject the testId into
their email address is kind of fallacious.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>

* Preventing upsert on Type is too greedy.

Turns out, we still want to upsert when Guests, because sometimes we
specify a particular Guest. This makes it so Guests who do not have an
email specified are treated like guest@example.com.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>

* Send new members to the Space#show upon RSVP completion

Also adds some text to say where the one-time passcode went

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>

* Blueprints clobber publicity and access level

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>

* Make the space explicit in another scenario

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>

Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>