Don't navigate to check who is signed in
There were a number of strange errors caused when navigating between
pages unnecessarily. This pulls the email address of the logged in user
to the header of the page in a data element, so we don't need to browse
around.

This also fixes a missing return that was causing strange test behavior.

Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>