Try to terse up README and make it more concrete
In [our Q1 2022
Retro](https://github.com/zinc-collective/convene/discussions/737), we
discussed not being clear on the Product, and how that gets in the way
of figuring out what to do in our ensemble sessions.

We took an edit pass on the README, making it more clear that it's an
operating system; and trying to integrate Zee's weird hand wavey
ramble-rants into something that people other than Zee could understand.

Co-authored-by: Kelly Hong <KellyAH@users.noreply.github.com>
Co-authored-by: Neer Malathapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Naomi Quinones <naomiquinones@users.noreply.github.com>