Marketplace: Products Controller persists Products!
This starts to shift our Furniture and Item system away from an
inheritance oriented rails-adjacent structure to a more Rails-forward
structure.

Furniture now places it's routes against the *room*, which allows it to
support either Singleton-type Resource routes, or more traditional
Resources routes.

Further, it brings the `Placeable` interface a bit closer to
ActiveRecord; which should reduce pain-points as we unwind the current
overly-restrictive and complex Furniture/Item design.