Marketplace: Add Products to Product Table via TurboStreams
This makes the user experience for adding a product even tighter; by
hiding the Product form until the Distributor is adding a new product.

Once they do, we replace the entire marketplace with the New Product
form; and upon successful create, we replace the entire marketplace with
the full Marketplace view!

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>