Marketplace 🛒: Member may add Products to Marketplace (#844)
* Add scaffolding for new Marketplace furniture.

Co-authored-by: Sadie Jay <19538219+sadiejay@users.noreply.github.com>
Co-authored-by: Zee <zspencer@users.noreply.github.com>

* Replace money gem with money-rails.

Co-authored-by: Sadie <19538219+sadiejay@users.noreply.github.com>
Co-authored-by: Zee <zspencer@users.noreply.github.com>

* Add Marketplace::Products model, table (not Items!), form and controller.

Co-authored-by: Sadie <19538219+sadiejay@users.noreply.github.com>
Co-authored-by: Zee <zspencer@users.noreply.github.com>

* Marketplace: Products Controller persists Products! (#856)

This starts to shift our Furniture and Item system away from an
inheritance oriented rails-adjacent structure to a more Rails-forward
structure.

Furniture now places it's routes against the *room*, which allows it to
support either Singleton-type Resource routes, or more traditional
Resources routes.

Further, it brings the `Placeable` interface a bit closer to
ActiveRecord; which should reduce pain-points as we unwind the current
overly-restrictive and complex Furniture/Item design.

* Marketplace: Products are shown!

* Use correct parameter name in request spec

Rails assumes that your parameter names will match the namespace of the
model. For example, a `Marketplace::Product#name` Rails will make HTTP Params
of `marketplace_product[name]`, I had previously assumed it would name
them `product[name]` because I thought it would discard the stuff before
the `::`.

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

* Rename from `routing_context` to `router`

@AnaUlin had suggested that `routing_context` was confusing; and I
agree. It's just the router! So let's name it that way!

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

* All `Placeable` now implement `to_partial_path`

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

Co-authored-by: Zee Spencer <zspencer@users.noreply.github.com>
Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

* Fix method lookup in class ancestry

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

* Marketplace: Present Products in a much cleaner way

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

* Adds Button to add Products to Marketplace

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

* Marketplace: Add Products to Product Table via TurboStreams

This makes the user experience for adding a product even tighter; by
hiding the Product form until the Distributor is adding a new product.

Once they do, we replace the entire marketplace with the New Product
form; and upon successful create, we replace the entire marketplace with
the full Marketplace view!

Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>

* Fixed test by redirecting to room rather than marketplace

Still need to figure out how to redirect to furniture

Co-authored-by:  Kelly <KellyAH@users.noreply.github.com>
Co-authored-by:  Zee <zspencer@users.noreply.github.com>
Co-authored-by:  Ana <anaulin@users.noreply.github.com>

* Marketplace: Validate product names

Validate that `Marketplace::Product` belongs to a space

Co-authored-by:  Kelly <KellyAH@users.noreply.github.com>
Co-authored-by:  Zee <zspencer@users.noreply.github.com>
Co-authored-by:  Ana <anaulin@users.noreply.github.com>

Co-authored-by: Sadie Jay <19538219+sadiejay@users.noreply.github.com>
Co-authored-by: Zee <zspencer@users.noreply.github.com>
Co-authored-by: Zee <50284+zspencer@users.noreply.github.com>
Co-authored-by: Neer Thapa <nirmalathapa@users.noreply.github.com>
Co-authored-by: Kelly <KellyAH@users.noreply.github.com>