Furniture: remove default Turbo Frame
This was added to try and make it easier to make dynamic websites,
however turbo-frames are really not the best use-case for this and we
are leaning more fully into turbo-streams and regular html.

I am hopeful this will also fix the weird display bug with the
Marketplace on piikup-marketplace.convene.zinc.coop as well.