Stripe: Sprout Initial Stripe Utility
https://github.com/zinc-collective/convene/issues/831
https://github.com/zinc-collective/convene/issues/1002

This is mostly because I want to be able to move forward with connecting
Vendor and Distributor accounts in the Marketplace, and to get started
fiddling with the Stripe Connect bits.

This also hopefully helps set the stage to getting rid of `Utility`
classes, and instead have those inherit fro `UtilityHookup` (kind of
like how new `Furniture` inherits from `FurniturePlacement`

To be honest, I should probably either do a rename pass to more clearly
indicate the old-way vs now-and-forward way of defining furniture, or do
a final cleanup pass on both of those before starting this, but this was
the nearest-mess-to-muck-with ( not clean... just.. uhhh... muck with)

Anyway!