Update CheckoutPolicy to match spec expectations (#1044)
This PR updates the CheckoutPolicy implementation to satisfy the expectations in the spec. I do not know if this is precisely the behavior we want -- e.g. maybe we want a "marketplace admin" (space owner?) to be able to always see a Checkout -- but we can expand the behavior later.