Marketplace: Differentiate between Checkout and Order
Where `Checkout` is useful for handling a Checkout flow for the
Payment Processor, it doesn't make a ton of sense from a Buyer
perspective once the Cart has been Checked out.

I've sprouted an `Order` model, with corresponding Controllers, and
moved the `checkouts/show` and `checkouts/_checkout` views to
`orders/show` and `orders/_order`.