Neighborhood: Support Cloudflare SSL w/Heroku
See:
- https://developers.cloudflare.com/ssl/edge-certificates/encrypt-visitor-traffic/
- https://support.cloudflare.com/hc/en-us/articles/205893698-Configure-Cloudflare-and-Heroku-over-HTTPS

So, apparently if you force ssl when you're using Cloudflare on top it
can put the application into a redirect loop.

This allows a Neighborhood who relies on Cloudflare for DNS + Security
to disable Rails' built in SSL redirection by setting the `FORCE_SSL`
environment variable to false.

Note: I would really prefer not to be using Cloudflare at all, however
they continue to provide the best free plan and we can't afford to pay
for a CDN at this time.